+++ src/main/java/com/iab/gdpr/exception/VendorConsentCreateException.java
@@ -1,9 +1,9 @@
 package com.iab.gdpr.exception;
 
+/**
+ * Exception for the case where consent string cannot be created
+ */
 public class VendorConsentCreateException extends VendorConsentException {
-    public VendorConsentCreateException(String message, Throwable cause) {
-        super(message, cause);
-    }
 
     public VendorConsentCreateException(String message) {
         super(message);