+++ src/main/java/com/iab/gdpr/exception/VendorConsentParseException.java
@@ -1,9 +1,9 @@
 package com.iab.gdpr.exception;
 
+/**
+ * Exception for the case where consent string cannot be parsed
+ */
 public class VendorConsentParseException extends VendorConsentException {
-    public VendorConsentParseException(String message, Throwable cause) {
-        super(message, cause);
-    }
 
     public VendorConsentParseException(String message) {
         super(message);