+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+## [3.0.1] - 12-21-2018
+
+### Changed
+- Removed VendorConsent class that was deprecated since version 2.0.1
+- Better handle index out of bounds conditions with corrupt consent strings
 
 ## [2.0.2] - 06-19-2018